requirements.txt
absl-py==0.9.0
astor==0.8.1
bleach==1.5.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
gast==0.3.3
grpcio==1.26.0
h5py==2.10.0
html5lib==0.9999999
idna==2.8
joblib==0.14.1
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
numpy==1.16.2
opencv-python==4.1.2.30
Pillow==7.0.0
protobuf==3.11.2
psutil==5.6.7
pyparsing==2.4.6
python-dateutil==2.8.1
requests==2.22.0
scikit-learn==0.22.1
scipy==1.2.1
six==1.14.0
tensorboard==1.7.0
tensorflow-gpu==1.7.0
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==0.16.0
cx_oracle